Refactor BufferPools to use a common template base class
Extracted duplicated buffer pooling logic from EnhancedBufferPool and
EnhancedAudioBufferPool into a new EnhancedTemplateBufferPool<T>
base class. This centralizes memory pressure handling, usage
statistics tracking, and cleanup routines, significantly improving
maintainability. Backward compatibility for the original APIs is
maintained via wrapper methods.
